fix issue #2 - don't download dataset if previously downloaded

+++ 2-sql-internals/entrypoint.sh
@@ -7,12 +7,21 @@ if [ ! -s "$PGDATA/PG_VERSION" ]; then
   apt-get update
   apt-get -yy install wget
 
-  echo "Downloading dataset.."
-  wget -O /docker-entrypoint-initdb.d/oz.sql.gz http://otvorenezmluvy.sk/data/otvorenezmluvy_dump_2013-09-08.sql.gz
+  if [[ ! -f "/docker-entrypoint-initdb.d/oz.sql.gz" ]]; then
+    echo "Downloading dataset.."
+    wget -O /docker-entrypoint-initdb.d/oz.sql.gz http://otvorenezmluvy.sk/data/otvorenezmluvy_dump_2013-09-08.sql.gz
+  else
+    echo "Using previously downloaded dataset: /docker-entrypoint-initdb.d/oz.sql.gz"
+  fi
 
-  echo "Unpacking.."
-  gunzip /docker-entrypoint-initdb.d/oz.sql.gz
+  if [[ ! -f "/docker-entrypoint-initdb.d/oz.sql" ]]; then
+    echo "Unpacking.."
+    gunzip /docker-entrypoint-initdb.d/oz.sql.gz
+  else
+    echo "Using previously unpacked dataset: /docker-entrypoint-initdb.d/oz.sql"
+  fi
 fi
 
 echo "Launching PostgreSQL"
 /docker-entrypoint.sh postgres
+
